+++ pentobi/qml/GameViewMobile.qml
@@ -11,6 +11,11 @@ import Qt.labs.settings 1.0
 import "." as Pentobi
 import "GameView.js" as Logic
 
+/** Game view optimized for mobile devices.
+    Landscape orientation is still experimentell because of bugs in Qt (with
+    Qt 5.12, there is sometimes a 1-pixel wide white line visible after
+    rotating the screen. To enable support for landscape, remove
+    android:screenOrientation="portrait" from AndroidManifest.xml. */
 Item
 {
     id: root